requirements.txt
faiss-cpu
numpy
pandas
pyarrow
scikit-learn
scipy
streamlit
torch
tqdm